CameraFollow does not follow based on jumps
CameraFollow no longer follows during jumps, but does re-track to when the player moves to a different floor as long as the Y limit is set. Can also be used to have the camera follow up/down staircases, as it will jump up to meet the player if they get too far away, but is not a required feature to use.

Set the cameraYLimit variable to less than how far the player will be moved in the Y direction when teleported. It will need to be relative to the camera, but it should not be hard to test and find out. Camera will otherwise remain on its Y axis that it's given on start. Offset can still be used to maintain a proper height difference between player and camera.